Refactor Song Comments so that CommentsList automatically updates after a new comment is added: SongLayout renders SongComments, passing it songId as a prop. SongComments renders CommentsContainer, passing it 'addSongComment' as 'addComment', the song entity as 'commentable', 'commentableId', and 'commentableType' as props. CommentsContainer imports the stylesheet for comments. CommentsContainer renders AddCommentForm, passing it 'addComment', 'commentableId', and 'commentableType' as props. CommentsContainer also renders CommentsList, passing it 'commentable'.
The idea is that SongComments does all the comments work that's specific to Songs (ie selecting the song entity and importing addSongComment).
This allows all of the comments components (CommentsContainer, AddCommentForm, CommentsList) to be entity-agnostic.
This is why we thread the addSongComment asyncThunk all the way down to AddCommentForm, and why we pass an entity object ('commentable') to CommentsList instead of an id -- we don't want any of the comments components to have to care which entity they are displaying comments for.